+++ CHANGELOG.md
@@ -2,7 +2,25 @@
 Change Log
 ==========================
 
-4.0.0 (unreleased)
+4.0.0-beta.2 (444)
+--------------------------
+
+### Improvements
+
+- Round the corners of current line highlight.
+- [beta][Big Sur] Match the inspector background to desktop color.
+- [beta][tirivial] Update titlebar color.
+
+### Fixes
+
+- [beta] Fix an issue in the navigation bar that the split editor button did not update when changing the split orientation.
+- [beta] Fix an issue in the appearance pane and the multiple replacement panel that the newly added setting was not selected.
+- [beta] Address an issue that the status bar rarely compressed vertically.
+- [beta] Workaround an issue that the initial text color of the pop-up menus in the status bar dimmed.
+
+
+
+4.0.0-beta (442)
 --------------------------
 
 ### New Features